Rename the matomo analytics package directory to be piwik
Also altered tracking hooks.
